+++ CHANGELOG.md
@@ -16,8 +16,11 @@ This project adheres to [Semantic Versioning 2.0.0](https://semver.org/spec/v2.0
 
 ### Fixed
 
+- Reminders create and update: calculate the next occurrence from now for recurring events to setup the correct reminder time
+
 ### Security
 
+- updated happy-DOM dependency to fix RCE attacks issue (https://github.com/advisories/GHSA-37j7-fg3j-429f)
 
 ---